api_topology/requirements.txt
python-Levenshtein>=0.21.0
neo4j>=5.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
